Make our services be what they are: type classes
Our endpoint classes require that we have an instance of the Effect type class
for F. Our services are also parameterized by that F type parameter, and they
must be the same F for our endpoints to work. There's not much difference
between the requirement for the Effect instance and the requirement for the
Service instances required by the Endpoint classes. So we can treat the
Services as type classes!